+++ src/model2/account.ts
@@ -1,4 +1,4 @@
-import { AccountFlags, AccountThresholds } from "./index";
+import { AccountFlags, AccountFlagsFactory, AccountThresholds } from "./index";
 
 export interface IAccountBase {
   id: string;
@@ -38,7 +38,7 @@ export class Account implements IAccount {
       homeDomain: row.homedomain,
       lastModified: row.lastmodified,
       thresholds: AccountThresholds.fromValue(row.thresholds),
-      flags: AccountFlags.fromValue(row.flags)
+      flags: AccountFlagsFactory.fromValue(row.flags)
     };
 
     return new Account(data);
