+++ src/model2/account_flags.ts
@@ -1,5 +1,3 @@
-import stellar from "stellar-base";
-
 export interface IAccountFlags {
   authRequired: boolean;
   authRevokable: boolean;
@@ -7,18 +5,6 @@ export interface IAccountFlags {
 }
 
 export class AccountFlags implements IAccountFlags {
-  public static fromValue(value: number): AccountFlags {
-    const fl = stellar.xdr.AccountFlags;
-
-    const data: IAccountFlags = {
-      authRequired: (value & fl.authRequiredFlag().value) > 0,
-      authRevokable: (value & fl.authRevocableFlag().value) > 0,
-      authImmutable: (value & fl.authImmutableFlag().value) > 0
-    };
-
-    return new AccountFlags(data);
-  }
-
   public authRequired: boolean;
   public authRevokable: boolean;
   public authImmutable: boolean;
