+++ src/model2/factories/index.ts
@@ -1,0 +1,1 @@
+export * from "./account_flags_factory";
